+++ ext/opentelemetry-ext-flask/setup.cfg
@@ -13,13 +13,13 @@
 # limitations under the License.
 #
 [metadata]
-name = opentelemetry-ext-wsgi
-description = WSGI Middleware for OpenTelemetry
+name = opentelemetry-ext-flask
+description = Flask tracing for OpenTelemetry (based on opentelemetry-ext-wsgi)
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 author = OpenTelemetry Authors
 author_email = cncf-opentelemetry-contributors@lists.cncf.io
-url = https://github.com/open-telemetry/opentelemetry-python/ext/opentelemetry-ext-wsgi
+url = https://github.com/open-telemetry/opentelemetry-python/ext/opentelemetry-ext-flask
 platforms = any
 license = Apache-2.0
 classifiers =
@@ -39,7 +39,7 @@ package_dir=
     =src
 packages=find_namespace:
 install_requires =
-    opentelemetry-api
+    opentelemetry-ext-wsgi
 
 [options.extras_require]
 test = flask~=1.0
