+++ ext/opentelemetry-ext-wsgi/setup.cfg
@@ -41,8 +41,5 @@ packages=find_namespace:
 install_requires =
     opentelemetry-api
 
-[options.extras_require]
-test = flask~=1.0
-
 [options.packages.find]
 where = src
