+++ ext/test-util/README.txt
@@ -1,0 +1,3 @@
+This (non-installable) directory should be added to the PYTHONPATH of any unit tests that require it.
+
+It contains shared uitlities for tests.
