+++ tox.ini
@@ -2,7 +2,7 @@
 skipsdist = True
 skip_missing_interpreters = True
 envlist =
-    py3{4,5,6,7,8}-test-{api,sdk,example-app,ext-wsgi,ext-http-requests,ext-jaeger}
+    py3{4,5,6,7,8}-test-{api,sdk,example-app,ext-wsgi,ext-flask,ext-http-requests,ext-jaeger}
     pypy3-test-{api,sdk,example-app,ext-wsgi,ext-http-requests,ext-jaeger}
     lint
     py37-{mypy,mypyinstalled}
@@ -18,6 +18,7 @@ deps =
 
 setenv =
   mypy: MYPYPATH={toxinidir}/opentelemetry-api/src/
+  PYTHONPATH={toxinidir}/ext/test-util
 
 changedir =
   test-api: opentelemetry-api/tests
@@ -25,6 +26,7 @@ changedir =
   test-ext-http-requests: ext/opentelemetry-ext-http-requests/tests
   test-ext-jaeger: ext/opentelemetry-ext-jaeger/tests
   test-ext-wsgi: ext/opentelemetry-ext-wsgi/tests
+  test-ext-flask: ext/opentelemetry-ext-flask/tests
   test-example-app: examples/opentelemetry-example-app/tests
 
 commands_pre =
@@ -35,9 +37,11 @@ commands_pre =
   example-app: pip install {toxinidir}/opentelemetry-sdk
   example-app: pip install {toxinidir}/ext/opentelemetry-ext-http-requests
   example-app: pip install {toxinidir}/ext/opentelemetry-ext-wsgi
+  example-app: pip install {toxinidir}/ext/opentelemetry-ext-flask
   example-app: pip install {toxinidir}/examples/opentelemetry-example-app
   ext: pip install {toxinidir}/opentelemetry-api
-  wsgi: pip install {toxinidir}/ext/opentelemetry-ext-wsgi[test]
+  wsgi,flask: pip install {toxinidir}/ext/opentelemetry-ext-wsgi
+  flask: pip install {toxinidir}/ext/opentelemetry-ext-flask[test]
   http-requests: pip install {toxinidir}/ext/opentelemetry-ext-http-requests
   jaeger: pip install {toxinidir}/opentelemetry-sdk
   jaeger: pip install {toxinidir}/ext/opentelemetry-ext-jaeger
@@ -72,7 +76,8 @@ commands_pre =
   pip install -e {toxinidir}/ext/opentelemetry-ext-azure-monitor
   pip install -e {toxinidir}/ext/opentelemetry-ext-http-requests
   pip install -e {toxinidir}/ext/opentelemetry-ext-jaeger
-  pip install -e {toxinidir}/ext/opentelemetry-ext-wsgi[test]
+  pip install -e {toxinidir}/ext/opentelemetry-ext-wsgi
+  pip install -e {toxinidir}/ext/opentelemetry-ext-flask[test]
   pip install -e {toxinidir}/examples/opentelemetry-example-app
 
 commands =
@@ -88,18 +93,19 @@ commands =
          ext/opentelemetry-ext-http-requests/src/opentelemetry \
          ext/opentelemetry-ext-jaeger/src/opentelemetry \
          ext/opentelemetry-ext-wsgi/src/opentelemetry  \
-         examples/opentelemetry-example-app/src/opentelemetry_example_app \
-         examples/opentelemetry-example-app/tests
+         ext/opentelemetry-ext-flask/src/opentelemetry  \
+         examples/opentelemetry-example-app/src/opentelemetry_example_app
 
-; The tests need extra pylint invocations, otherwise we get name clashes within the "tests" modules
-; (symptom: imports from test don't work)
-
+; The tests need extra pylint invocations, otherwise we get name clashes within the "tests" packages
+; (symptom: imports from tests don't work)
+  pylint examples/opentelemetry-example-app/tests
   pylint opentelemetry-sdk/tests
   pylint opentelemetry-api/tests
   pylint ext/opentelemetry-ext-azure-monitor/tests
   pylint ext/opentelemetry-ext-http-requests/tests
   pylint ext/opentelemetry-ext-jaeger/tests
   pylint ext/opentelemetry-ext-wsgi/tests
+  pylint ext/opentelemetry-ext-flask/tests
 
   flake8 .
   isort --recursive .
